fix: resolve silent failures, float precision, and error handling issues
Fixes from codebase architecture audit:

- Fix BigInt array access without bounds checking in ledger_cursor.rs
  (could panic on zero values from to_u64_digits())

- Replace float money calculations with integer arithmetic in:
  - payment_release_service.rs (release amount calculation)
  - database/contracts.rs (refund calculation)
  - payment_release_service/tests.rs (match production)
  This prevents precision loss in financial calculations.

- Replace .unwrap_or_default() with error-preserving fallback in HTTP
  response body extraction across 30+ locations:
  - chatwoot/client.rs (26 instances)
  - support_bot/embeddings.rs
  - notifications/telegram.rs
  - support_bot/llm.rs
  Now shows "<failed to read body: {error}>" instead of empty string.

- Standardize email domain defaults to noreply@decent-cloud.org
  (was inconsistent: decloud.org vs decent-cloud.org)

- Document additional architectural issues in TODO.md:
  - Geographic region mapping duplicate (200+ lines)
  - CLI commands with todo!() stubs
  - Hardcoded token value ($1 USD)
  - Hardcoded secrets in version control
  - Half-implemented canister proxy
  - Missing api-server doctor checks
  - Manual setup steps that could be automated

+++ api/src/database/contracts.rs
@@ -1043,9 +1043,9 @@ impl Database {
             return 0;
         }
 
-        // Calculate prorated refund: (time_remaining / total_duration) * payment_amount
-        let refund_amount = (payment_amount_e9s as f64 * time_remaining_ns as f64
-            / total_duration_ns as f64) as i64;
+        // Calculate prorated refund using integer arithmetic (avoid float precision loss)
+        let refund_amount = ((payment_amount_e9s as i128) * (time_remaining_ns as i128)
+            / (total_duration_ns as i128)) as i64;
 
         // Ensure non-negative
         refund_amount.max(0)
